requirements.txt
Django==2.1.1
Pillow==5.3.0
pytz==2018.5
SQLAlchemy==1.2.11
pip==18.1
setuptools==40.4.3
wheel==0.32.1